inicio testes de carro e mudanca do tipo de preco aluguel em carro, de float para double2

+++ src/Entidades/Carro.java
@@ -61,7 +61,7 @@ public class Carro {
 		this.ano = ano;
 	}
 
-	public float getPrecoAluguel() {
+	public double getPrecoAluguel() {
 		return precoAluguel;
 	}
